Implement sqlserver support for most tests
When swapping databases, we cannot delete and recreate because we would run in trouble with filenames. `ALTER DATABASE x MODIFY NAME = y` does not rename the underlying file and thus a subsequent `CREATE DATABASE x` would fail. Thus for mssql we create both the transaction and the final database from the start and then only clear schemas instead of drop + create.