Also transform lowercase table names to uppercase DB2 names in table hooks.
This might not be done in all places and it might be done more generalized as a dialect specific name mangling system. We only fixed cases that caused trouble in a concrete usage scenario. We are a bit dependent here on internal mechanics of SQLAlchemy so it is hard to predict where exactly those replacements are needed.

The concrete problem we had here, if we create an all-lowercase table in DB2, we even can't drop it in cleanup anymore since we think SQLAlchemy just converted a case-insensitive all-caps table name into an all-lowercase one and we do the reverse.