try sequentializing python version tests
The ibm_db2 service is only started once for the whole matrix execution with steps for different python versions. Due to the artificially small tests, this leads to actions that DB2 consideres as deadlock even though zookeeper is used to ensure that any stage is only touched by exactly one process at any time.

It could be that using wildly different graphs on the same instance_id might cause this, but this problem will not affect any realistic usecases of pipedag.