wrap another operation in retry for DB2 deadlock victim eviction
Surely the code should deduplicate retry code. And it should be better understood for DB2 whether it is possible to avoid transaction collisions in system schema tables. The problem is caused since we like to transfer tables to temporary schema in parallel to working on this schema. In theory this should not be a problem because we work on different tables. In practice this is more complicated since we do column type reflection and modify tables managed by the same system table.