Implemented Trace Hook functionality for debugging caching problems.
The idea is that a derived class of TraceHook may be given to Flow.run() which is informed about all sorts of events during execution. It would need to take care of race conditions and multi-node execution (see Dask engine) internally but should have all inputs needed to do so. It could piggy-back on RunContextServer for client-server communication even though the interface would still have to be designed for this.

For now, there is just a trivial PrintTraceHook implementation which prints activity on debug logging level. Don't forget to set environment variable DEBUG=1 to see it.